# Improve filbuf performance