--- a/.appveyor.yml
+++ b/.appveyor.yml
@@ -134,7 +134,7 @@ environment:
       APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2022
       CXXSTD: 17
       # This test sometimes fails on recent images when using CMake
-      CXX_FLAGS: -DNOWIDE_DISABLE_CIN_TEST=ON
+      CMAKE_FLAGS: -DBOOST_NOWIDE_DISABLE_CIN_TEST=ON
       GENERATOR: Visual Studio 17 2022
       configuration: Debug
       BOOST_ROOT: C:\Libraries\boost_1_77_0
@@ -175,7 +175,7 @@ for:
       - if exist %INSTALL_DIR%\ (rmdir /S /Q %INSTALL_DIR%)
       - mkdir __build_cmake_test__
       - cd __build_cmake_test__
-      - cmake -G "%GENERATOR%" -DCMAKE_CXX_STANDARD=%CXXSTD% %CXX_FLAGS% -DCMAKE_INSTALL_PREFIX=%INSTALL_DIR% ..
+      - cmake -G "%GENERATOR%" -DCMAKE_CXX_STANDARD=%CXXSTD% %CMAKE_FLAGS% -DCMAKE_INSTALL_PREFIX=%INSTALL_DIR% ..
     build_script: cmake --build . --config %configuration% --parallel 4
     test_script:
       - ctest --output-on-failure -C %configuration% --parallel 4
@@ -186,7 +186,7 @@ for:
       # Build consumer example test
       - cmake --build . --config %configuration% --target install
       - del /F /S /Q *
-      - cmake -DBOOST_NOWIDE_INSTALL_TEST=ON -G "%GENERATOR%" -DCMAKE_CXX_STANDARD=%CXXSTD% %CXX_FLAGS% -DCMAKE_PREFIX_PATH=%APPVEYOR_BUILD_FOLDER%\installed ../test/cmake_test
+      - cmake -DBOOST_NOWIDE_INSTALL_TEST=ON -G "%GENERATOR%" -DCMAKE_CXX_STANDARD=%CXXSTD% %CMAKE_FLAGS% -DCMAKE_PREFIX_PATH=%APPVEYOR_BUILD_FOLDER%\installed ../test/cmake_test
       - cmake --build . --config %configuration%
       - ctest --output-on-failure -C %configuration% --parallel 4